==> 10_preparation.qmd <==
# Preparation {.unnumbered}

All informations and documents needed for the FEFACCION 2025 preparation.

The link toward the workshop for the Friday is available here [Forest dynamic with R](workshop/workshop.html) .
