==> 11_traits.qmd <==
# Traits {.unnumbered}

We will measure the following functional traits.

-   **LA**: leaf area $cm^{2}$
-   **LDMC**: leaf dry matter content $g.g^{-1}$
-   **LT**: leaf thickness ($\mu m$)
-   **SLA**: specific leaf area ($g.cm^{-2}$)
-   **TLP**: leaf water potential at which leaf cells lose turgor ($MPa$), a key drought tolerance trait.
-   **Nmass**: leaf nitrogen mass content ($mg.g^{-1}$)
-   **Pmass**: leaf phosphorus mass content ($mg.g^{-1}$)
-   **WD**: wood density ($g.cm^{-3}$,)

They will necessitate the following raw measurements:

-   Leaf fresh thickness (LT)
-   Leaf SPAD measurement (CC)
-   Leaf fresh weight (LDMC, SLA, RWC)
-   Leaf scan (LA, MajVLA)
-   Leaf saturated weight (RWC)
-   Leaf saturated water potential (TLP)
-   Wood dry mass (WDM)
-   Wood saturated volume (WSV)
