==> 13_material.qmd <==
# Material {.unnumbered}

We need to gather the following material.

## Big material

-   Camera
-   Binoculars
-   Big Shot
-   Big Shot weight
-   Big Shot wire
-   Helmets (\~15)
-   Micrometer
-   SPAD
-   Oven
-   Vapor pressure osmometer (Vapro 5520, Wescor, Logan, UT)
-   Vapro software (Vapro Lab Report)
-   Liquid nitrogen
-   Fridge
-   Distilled water
-   Precision balance
-   Scanner
-   Computer

## Small material

-   Field support
-   AA batteries
-   Flat batteries
-   Cisors
-   Ziplocks
-   Labels
-   Tissues
-   Bin bafs
-   Plastic box
-   Mister
-   Pencil
-   Markers
-   Maps
-   Field notebook
-   Pruning shears
-   Envelopes
-   Journal
-   Press
-   Paper towel
-   Metal tea ball
-   Tin foil
-   Needle
-   Liquid nitrogen gloves
-   Liquid nitrogen goggles
-   Liquid nitrogen contenant
-   2 Tweezers
-   Cork borer
