==> 14_maps.qmd <==
```{r set}
#| include: false
library(tidyverse)
library(sf)
species <- c(
 "ANDIROBA",
 "ANGELIM VERMELHO",
 "ARARAQUANGA",
 "CASTANHA DO PARA",
 "COPAIBA",
 "CUMARU",
 "CUPIUBA",
 "FAVA",
 "FAVA AMARGA",
 "FAVA ORELHA DE MACACO",
 "FREIJO",
 "JARANA",
 "JUTAI",
 "MARUPA",
 "MELANCIEIRA",
 "PARAPARA",
 "PIQUIA",
 "TACHI BRANCO",
 "TATAJUBA",
 "VIROLA CASCA DE VIDRO",
 "CEDRURANA",
 "QUARUBA"
 )
forest <- readxl::read_excel("data/raw/plantation/Dados brutos do inventário da estrada 5.xlsx") %>% 
  rename(plate = Placa, name = `Nome vernacular`,
         x = `Easting (UTM)`, y =  `Northing (UTM)`) %>% 
  mutate(dbh = as.numeric(CAP1)/pi) %>% 
  select(plate, name, x, y, dbh) %>% 
  filter(!is.na(name)) %>% 
  filter(name %in% species) %>% 
  group_by(name) %>% 
  top_n(10, dbh) %>% 
  mutate(code = str_sub(gsub(" ", "", name), 1, 5)) %>% 
  mutate(label = paste0(code, "\n", plate))
roads <- read_sf("data/raw/plantation/BosquePesquisaEmbrapa.gpkg", "EstradasdoPPesquisa")
inventory <- read_sf("data/raw/plantation/BosquePesquisaEmbrapa.gpkg", "Inventario")
mother <- read_sf("data/raw/plantation/BosquePesquisaEmbrapa.gpkg", "MATRIZES")
limits <- read_sf("data/raw/plantation/BosquePesquisaEmbrapa.gpkg", "PerimetroEstrada5")
path <- read_sf("data/raw/plantation/BosquePesquisaEmbrapa.gpkg", "TrilhaParqPesq")
inventory <- forest %>% 
  st_as_sf(coords = c("x", "y"), crs = 31981)
```

# Maps {.unnumbered}

We will collect the following species:

```{r splist}
#| message: false
#| warning: false
forest %>% 
  group_by(name, code) %>% 
  summarise(N = n()) %>% 
  arrange(desc(N)) %>% 
  knitr::kable()
```

We selected the 10 biggest individuals per species resulting in the following maps of individuals to be collected:

```{r map1}
#| message: false
#| warning: false
ggplot() +
  geom_sf(data = roads) +
  geom_sf(data = path) +
  geom_sf(data = inventory, aes(col = code)) +
  theme_bw()
```

```{r map2}
#| message: false
#| warning: false
g <- ggplot() +
  geom_sf(data = path, col = "darkgrey") +
  geom_sf(data = inventory) +
  ggrepel::geom_text_repel(
      data = inventory,
      aes(label = label, geometry = geometry),
      stat = "sf_coordinates",
      size = 2, max.overlaps = 20) +
  theme_bw() +
  theme(axis.title = element_blank(), axis.text = element_blank(),
        axis.ticks = element_blank(), axis.line = element_blank())
ggsave(g, file = "documents/map.pdf",
       width = 420, height = 297, unit = 'mm', dpi = 300)
knitr::include_graphics("documents/map.pdf")
```
