==> 22_scans.qmd <==
# Scans {.unnumbered}

```{r set}
#| include: false
library(tidyverse)
```

We first use the following bash code to aggregate images and convert pdf to png.

```         
cd data/derived/
mkdir pdfs
ls ../raw/scans/*/*
cp ../raw/scans/*/* pdfs
mkdirs pngs
for file in $(ls pdfs/)
do
  pdftoppm pdfs/$file pngs/$file -png
done
```

We then use ImageJ (<https://imagej.net/ij/download.html>) with the following steps:

-   Open virtual stack
-   Type \> 8-Bit
-   Adjust \> Threshold \> Automatic
-   Analyze Particles
    -   Size: 1000-Infinity
    -   Circularity: 0-1
    -   Display results, clear results, exclude on edges, summarize, composite ROIs
    -   with or without include holes depending on the question

The intermediate threshold images with detected are saved for manual check.

```{r data}
#| eval: false
la_ex <- read_csv("data/derived/area_excluded_summary.csv") %>% 
  select(Slice, `Total Area`) %>% 
  rename(slice = Slice, area_excluded = `Total Area`) %>% 
  separate(slice, c("vernacular", "tree", "leaf")) %>% 
  mutate(area_excluded = area_excluded / (1275*1755) * (21*29.7))
la_in <- read_csv("data/derived/area_included_summary.csv") %>% 
  select(Slice, `Total Area`) %>% 
  rename(slice = Slice, area_included = `Total Area`) %>% 
  separate(slice, c("vernacular", "tree", "leaf")) %>% 
  mutate(area_included = area_included / (1275*1755) * (21*29.7))
la_ex %>% 
  left_join(la_in) %>% 
  write_tsv("data/derived/la.tsv")
```

```{r figure}
#| message: false
#| warning: false
read_tsv("data/derived/la.tsv") %>% 
  group_by(vernacular) %>% 
  mutate(la_s = mean(area_included)) %>% 
  ungroup() %>% 
  mutate(vernacular = fct_reorder(vernacular, la_s)) %>% 
  ggplot(aes(vernacular,
             area_included)) +
  geom_boxplot() +
  geom_jitter(aes(col = as.factor(tree)),
              width = 0.2, size = 3) + 
  theme_bw() +
  labs(x = "Vernacular name", y = "LA [ cm2 ]") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1),
        legend.position = "none") +
  scale_y_log10()
```

```{r problems}
#| message: false
#| warning: false
read_tsv("data/derived/la.tsv") %>% 
  filter(area_included < 1) %>% 
  knitr::kable()
```

> We have issues with ANGEL 5677 leaves 2 and 4 because scans are touching the border. I have no better solution than manually adding a black line in a photo editor to treat them beforehand.

## Manual check

Masks generated with the option allowing holes to be filled were manually inspected for (*NA* when no existing mask):

-   *missing_part*: missing part in the leave due to herbivory or other damage: *0* no missing part, *1* moderate missing part, and *2* big missing part **=\> code 1 can probably be included, but code 2 should probably be remove for LA**

-   *anomaly*: any anomaly in the generated masks:

    -   *ok*: no problem

    -   *abn_area*: abnormal area are masked, detected because they don't have the shape of a leaf. Looking at the corresponding scan, it is because these areas correspond to some spaces between the folioles that are fully surrounded by leaf part. So they are interpreted as holes and filled. **=\> For these individuals (or maybe species), consider the unfilled area even for the LA**. TO DO if you agree

    -   *fol_touch*: this is the same problem than *abn_area*, except for the FAVA for which the foliolules are actually touching and form a compact area on the mask.**=\> For these individuals (or maybe species), consider the unfilled area even for the LA** TO DO if you agree

    -   *miss_petiolul*: for double-compound leaves, the petiolule is not included, when it should have been
    
    -   *petiol_included*: the petiole has been included, when it should not have

    -   *one_fol*: just one foliole when the species has compound leaves

-   *touching*: *0* no part touching, *1* leaf touching leading to part of the leaf not taken into account, *2* leaf touching leading to all the leaf not taken into account **=\> TO DO (Sylvain?) to be corrected by adding a black line with a photo editor?**

Raw leaf scans are manually checked for the following problems, but this is highly subjective:

-   *ok*: no problem (or very small area concerned)

-   *damaged*: part of the leaf or the whole leaf appears black/brown or yellowish on the scan or there are any other damages (except missing part of the leaf) suggesting that the leaf is decaying or is attacked by insects

-   *young*: the bright green colour of the leave compared to the other leaves from the individual and/or the species suggest non mature leaves
